CDK-515: Moved spark tests to their own module.
