+++ kite-data/pom.xml
@@ -29,6 +29,7 @@
     <module>kite-data-hcatalog</module>
     <module>kite-data-hbase</module>
     <module>kite-data-mapreduce</module>
+    <module>kite-data-spark</module>
   </modules>
 
   <parent>
